[Refactor] 🍎Map / Direction 공통 함수 컴포넌트화
